Change the signature of the TwilioUsage __invoke magic method
This change is being made for several reasons:

1. Because the information supplied from the user, on the command-line
will be a string, not a date, and the command should, in my opinion, not
be responsible for preparing the input of another process, rather just
handing the input it receives to that command, which will handle
preparing/filtering/vetting that input.
2. To add the ability to set a limit for the number of records returned
from requests to the API.

Update the call to the TwilioUsage __invoke magic method to bring it in
line with the change to the method's signature.